Fixing cyclic dependencies between algorithm, segmented_algorithm, and parallel_executor modules
